LayoutTests/fast/webgpu/regression/repro_290328-expected.txt

CONSOLE MESSAGE: Pass: no validation error
